- make rest handler code as simple as possible, defer vendor logic to graph (api) package - introduce graph/api as the high level api
